refactor(abw): decompose listicle writer
